requirements.txt
Flask==1.1.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
mysql-connector-python==8.0.30
numpy==1.23.4
openpyxl==3.0.10
pandas==2.0.1
SQLAlchemy==1.4.22
virtualenv==20.16.5